Fix build with release log disabled https://bugs.webkit.org/show_bug.cgi?id=303741
Reviewed by Chris Dumez and Per Arne Vollan.

* Source/WebCore/Modules/mediastream/libwebrtc/LibWebRTCMediaEndpoint.cpp:
(WebCore::LibWebRTCMediaEndpoint::LibWebRTCMediaEndpoint):

peerConnection is only used for logging.

* Source/WebCore/html/HTMLMediaElement.cpp:
(WebCore::HTMLMediaElement::setNetworkState):
(WebCore::HTMLMediaElement::mediaPlayerCharacteristicChanged):

HTMLMEDIAELEMENT_RELEASE_LOG() has an empty body when release log is disabled.
Add a stub body to avoid -Werror=empty-body with one-line control clauses.

* Source/WebKit/NetworkProcess/storage/SQLiteStorageArea.cpp:
(WebKit::SQLiteStorageArea::allItems):
Explicitly cast StorageError enum to uint8_t for log parameter.

Canonical link: https://commits.webkit.org/304148@main

+++ Source/WebKit/NetworkProcess/storage/SQLiteStorageArea.cpp
@@ -334,7 +334,7 @@ HashMap<String, String> SQLiteStorageArea::allItems()
             if (result.has_value())
                 items.add(key, result.value());
             else {
-                RELEASE_LOG_ERROR(Storage, "SQLiteStorageArea::allItems failed during read from cache (%hhu)" PUBLIC_LOG_STRING, result.error());
+                RELEASE_LOG_ERROR(Storage, "SQLiteStorageArea::allItems failed during read from cache (%hhu)" PUBLIC_LOG_STRING, static_cast<uint8_t>(result.error()));
                 return { };
             }
         }
